refactor(*): tidy up the structure
