+++ src/App.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { Route } from "react-router-dom";
 import Login from "./app/views/login/login";
 import List from "./app/views/list/list";
-import Header from "./app/components/Header"
+import Header from "./app/components/Header/Header"
 import styles from "./App.module.scss";
 
 const App = () => {
