+++ src/app/views/diagram/rappid.scss
