+++ src/app/views/diagram/rappid.tsx
