+++ src/app/views/login/login.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoginForm from '../../views/login/components/login_form';
+import LoginForm from './components/LoginForm';
 
 const Login = () => {
     return <div className="login-page-container">
