+++ src/index.css
@@ -1,8 +1,5 @@
-@import 'app/components/shared_components.scss';
-
 body {
   margin: 0;
-  background-color: var(--secondary);
   font-family: 'Epilogue', -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
